+++ mmo-game-server-dotnet/MMOGameServer/Middleware/WebSocketMiddleware.cs
@@ -131,8 +131,8 @@ public class WebSocketMiddleware
             _logger.LogInformation($"Client {client.Player.UserId} intentional logout - skipping soft disconnect.");
             
             // Save player position - handle death state specially
-            float saveX = client.Player.X;
-            float saveY = client.Player.Y;
+            int saveX = client.Player.X;
+            int saveY = client.Player.Y;
             
             // If player is dead or awaiting respawn, save them at spawn point instead
             if (!client.Player.IsAlive || client.Player.IsAwaitingRespawn)
