+++ mmo-game-server-dotnet/MMOGameServer/Services/PlayerService.cs
@@ -187,7 +187,7 @@ public class PlayerService
     /// <summary>
     /// Update player position and handle chunk tracking.
     /// </summary>
-    private void UpdatePlayerPosition(Player player, float newX, float newY)
+    private void UpdatePlayerPosition(Player player, int newX, int newY)
     {
         // Update chunk tracking and visibility
         _terrainService.UpdatePlayerChunk(player, newX, newY);
